photos/scan_qr service tune refactoring
