+++ FRONTEND/fastparking/parking/services.py
@@ -95,8 +95,11 @@ def format_hours(hours: float) -> str:
 
 
 def format_duration(duration: float | None) -> str | None:
+    fmt = format_hours(duration)
+    if fmt:
+        fmt = " = " + fmt
     return (
-        f"{duration:.2f}h = {format_hours(duration)}" if duration is not None else None
+        f"{duration:.2f}h{fmt}" if duration is not None else None
     )
