+++ FRONTEND/fastparking/photos/services.py
